Quick push 17/10/2019
Added a goal that behaves like prevent objects and uses the same save system line the player to optimize code.
The Goal will later end the level when touched and the player will get a reward.
In addition I changed the safe syntax.
(Changed "[P]" to "[Player]")
Goal can also be accessed in the Level Editor to place it.

Will be working on futher optimizing the game and engine.
I also will be working on bringing back AABB colision or I´ll try to understand the current Physics library